update sample app
