+++ .gitignore
@@ -1,2 +1,4 @@
 .gradle/
 .idea/
+*.iml
+build/
