--- samples/simple/AndroidManifest.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.familygraph.android" 
-    android:versionCode="1"
-    android:versionName="1.0">
-    <application android:icon="@drawable/icon" 
-        android:label="@string/app_name"
-        android:debuggable="true">
-        <activity android:name=".Example" 
-            android:label="@string/app_name"
-            android:configChanges="keyboardHidden">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
-    </application>
-    <uses-permission android:name="android.permission.INTERNET"></uses-permission>
-    <uses-sdk android:minSdkVersion="3"/>
-</manifest>
