+++ samples/simple/app/src/main/java/com/familygraph/android/sample/SessionStore.java
@@ -1,0 +1,38 @@
+package com.familygraph.android.sample;
+
+
+import com.familygraph.android.FamilyGraph;
+
+import android.content.Context;
+import android.content.SharedPreferences;
+import android.content.SharedPreferences.Editor;
+
+public class SessionStore {
+
+    private static final String TOKEN = "access_token";
+    private static final String EXPIRES = "expires_in";
+    private static final String KEY = "familygraph-session";
+
+    public static boolean save(FamilyGraph session, Context context) {
+        Editor editor = context.getSharedPreferences(KEY, Context.MODE_PRIVATE)
+                .edit();
+        editor.putString(TOKEN, session.getAccessToken());
+        editor.putLong(EXPIRES, session.getAccessExpires());
+        return editor.commit();
+    }
+
+    public static boolean restore(FamilyGraph session, Context context) {
+        SharedPreferences savedSession = context.getSharedPreferences(KEY,
+                Context.MODE_PRIVATE);
+        session.setAccessToken(savedSession.getString(TOKEN, null));
+        session.setAccessExpires(savedSession.getLong(EXPIRES, 0));
+        return session.isSessionValid();
+    }
+
+    public static void clear(Context context) {
+        Editor editor = context.getSharedPreferences(KEY, Context.MODE_PRIVATE)
+                .edit();
+        editor.clear();
+        editor.commit();
+    }
+}
