--- samples/simple/default.properties
@@ -1,5 +1,0 @@
-# Indicates whether an apk should be generated for each density.
-split.density=false
-android.library.reference.1=../../familygraph
-# Project target.
-target=android-3
