--- samples/simple/src/com/familygraph/android/SessionStore.java
@@ -1,57 +1,0 @@
-/*
- * Copyright 2012 MyHeritage, Ltd.
- *
- * The MyHeritage Family Graph Android SDK Example is based on the Facebook
- * Android simple example SDK:
- * Copyright 2012 Facebook, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.familygraph.android;
-
-import com.familygraph.android.FamilyGraph;
-
-import android.content.Context;
-import android.content.SharedPreferences;
-import android.content.SharedPreferences.Editor;
-
-public class SessionStore {
-
-    private static final String TOKEN = "access_token";
-    private static final String EXPIRES = "expires_in";
-    private static final String KEY = "familygraph-session";
-
-    public static boolean save(FamilyGraph session, Context context) {
-        Editor editor = context.getSharedPreferences(KEY, Context.MODE_PRIVATE)
-                .edit();
-        editor.putString(TOKEN, session.getAccessToken());
-        editor.putLong(EXPIRES, session.getAccessExpires());
-        return editor.commit();
-    }
-
-    public static boolean restore(FamilyGraph session, Context context) {
-        SharedPreferences savedSession = context.getSharedPreferences(KEY,
-                Context.MODE_PRIVATE);
-        session.setAccessToken(savedSession.getString(TOKEN, null));
-        session.setAccessExpires(savedSession.getLong(EXPIRES, 0));
-        return session.isSessionValid();
-    }
-
-    public static void clear(Context context) {
-        Editor editor = context.getSharedPreferences(KEY, Context.MODE_PRIVATE)
-                .edit();
-        editor.clear();
-        editor.commit();
-    }
-}
